Fix a bunch of PVS-Studio warnings.
- Use fseeko()/ftello() instead of fseek()/ftell().
- Explicitly check memcmp() != 0.
- Reorganized a few structs to use less memory.
- Save pointers instead of calling functions repeatedly.
- Remove NULL checks for delete.

MessageWidget::paintEvent(): Reduce the scope of painter so it's only
constructed when it's needed.

QRvtHToolWindow::openRvtH(): Properly delete rvth_tmp instead of deleting
d->rvth again, which may result in a double-free.

QRvtHToolWindow::on_actionOpenDevice_triggered(): delete selectdeviceDialog
when we're done using it.

SelectDeviceDialog.hpp, DeviceQueryData(): Initialize size in the zero-arg
constructor.

libwiicrypto/cert.c, cert_fakesign_ticket(): Use size_t for signing_offset.

+++ src/libwiicrypto/cert.c
@@ -440,7 +440,7 @@ int cert_fakesign_ticket(uint8_t *ticket_u8, size_t size)
 	do {
 		// Calculate the SHA-1 of the ticket.
 		// If the first byte is 0, we're done.
-		static const unsigned int signing_offset = offsetof(RVL_Ticket, issuer);
+		static const size_t signing_offset = offsetof(RVL_Ticket, issuer);
 		sha1_update(&sha1, size - signing_offset, (const uint8_t*)(&ticket_u8[signing_offset]));
 		sha1_digest(&sha1, sizeof(digest), digest);
 	} while (digest[0] != 0 && ++(*fake) != 0);
